+++ internal/component/database_observability/mysql/collector/schema_table.go
@@ -97,7 +97,6 @@ const (
 
 type SchemaTableArguments struct {
 	DB              *sql.DB
-	InstanceKey     string
 	CollectInterval time.Duration
 	EntryHandler    loki.EntryHandler
 
@@ -110,7 +109,6 @@ type SchemaTableArguments struct {
 
 type SchemaTable struct {
 	dbConnection    *sql.DB
-	instanceKey     string
 	collectInterval time.Duration
 	entryHandler    loki.EntryHandler
 
@@ -169,7 +167,6 @@ type foreignKey struct {
 func NewSchemaTable(args SchemaTableArguments) (*SchemaTable, error) {
 	c := &SchemaTable{
 		dbConnection:    args.DB,
-		instanceKey:     args.InstanceKey,
 		collectInterval: args.CollectInterval,
 		entryHandler:    args.EntryHandler,
 		logger:          log.With(args.Logger, "collector", SchemaTableName),
@@ -249,7 +246,6 @@ func (c *SchemaTable) extractSchema(ctx context.Context) error {
 		c.entryHandler.Chan() <- database_observability.BuildLokiEntry(
 			logging.LevelInfo,
 			OP_SCHEMA_DETECTION,
-			c.instanceKey,
 			fmt.Sprintf(`schema="%s"`, schema),
 		)
 	}
@@ -294,7 +290,6 @@ func (c *SchemaTable) extractSchema(ctx context.Context) error {
 			c.entryHandler.Chan() <- database_observability.BuildLokiEntry(
 				logging.LevelInfo,
 				OP_TABLE_DETECTION,
-				c.instanceKey,
 				fmt.Sprintf(`schema="%s" table="%s"`, schema, tableName),
 			)
 		}
@@ -337,7 +332,6 @@ func (c *SchemaTable) extractSchema(ctx context.Context) error {
 		c.entryHandler.Chan() <- database_observability.BuildLokiEntry(
 			logging.LevelInfo,
 			OP_CREATE_STATEMENT,
-			c.instanceKey,
 			fmt.Sprintf(
 				`schema="%s" table="%s" create_statement="%s" table_spec="%s"`,
 				table.schema, table.tableName, table.b64CreateStmt, table.b64TableSpec,